broke out armies to its own component
