+++ Client/src/components/game/Province.js
@@ -1,4 +1,5 @@
 import './Province.css';  
+import Army from './Army.js'
 
 /**
  * Component for provinces in the main view. 
@@ -9,36 +10,6 @@ import './Province.css';
 
 export default function Province({ id, onProvinceClick, onArmyClick, name, owner, armies, moveArmy, session, player }) {
   // If start dragging an army  
-  function handleOnDrag(e, whatArmy, fromProvince, fromSlot){
-
-    e.target.classList.add('drag_army_start');
-    
-    //e.originalEvent.dataTranfser.setData("widgetType", widgetType);
-    e.dataTransfer.setData("whatArmy", whatArmy);
-    e.dataTransfer.setData("fromProvince", fromProvince);
-    e.dataTransfer.setData("fromSlot", fromSlot);
-  }
-  
-  // If dropping an army in a province
-  function handleOnDrop(e, toProvince){
-    const whatArmy = e.dataTransfer.getData("whatArmy");
-    const fromProvince   = e.dataTransfer.getData("fromProvince");
-    const fromSlot   = e.dataTransfer.getData("fromSlot");
-
-    e.target.style.animation = 'army_appear 1s';
-    console.log(e.target.style.animation);
-
-    moveArmy(fromProvince, toProvince, whatArmy, fromSlot);
-  }
-
-  // Stop dragging animation when dropping the element
-  const handleDragEnd = (e) => {
-    e.target.classList.remove('drag_army_start');
-  };
-  
-  function handleDragOver(e){
-      e.preventDefault();
-  }
 
   // Setup province colors and set color for each player
   const playerNames = session.slot_names;
@@ -51,22 +22,52 @@ export default function Province({ id, onProvinceClick, onArmyClick, name, owner
 
   // Set color of current province
   const color = playerColors[owner];
-  // If an army exists in a slot, show that it exists and make it draggable
-  const army1Exists = (armies[0] == null) ? 0.2 : 1.0;
-  const army2Exists = (armies[1] == null) ? 0.2 : 1.0;
-  const army3Exists = (armies[2] == null) ? 0.2 : 1.0;
-  const army4Exists = (armies[3] == null) ? 0.2 : 1.0;
-  const army1Draggable = (armies[0] == null || player.name != owner) ? false : true;
-  const army2Draggable = (armies[1] == null || player.name != owner) ? false : true;
-  const army3Draggable = (armies[2] == null || player.name != owner) ? false : true;
-  const army4Draggable = (armies[3] == null || player.name != owner) ? false : true;
+
+  /**
+   * @brief: Dropping an army on a province
+   * 
+   * @param {*} e: The HTML element of the army
+   * @param {*} toProvince: What province number it is being dropped in
+   */
+  function handleOnDrop(e, toProvince){
+    const whatArmy = e.dataTransfer.getData("whatArmy");
+    const fromProvince   = e.dataTransfer.getData("fromProvince");
+    const fromSlot   = e.dataTransfer.getData("fromSlot");
+
+    moveArmy(fromProvince, toProvince, whatArmy, fromSlot);
+  }
+  
+  function handleDragOver(e){
+      e.preventDefault();
+  }
   
   return (
       <div className='province' id={id} owner={owner} style={{background: color}} 
         onDragOver={handleDragOver} onDrop={(e) => {handleOnDrop(e, id)} }>
       
       <button className='province_name' onClick={onProvinceClick}>{name}</button>
+      <Army key={name + '_army1'} name='army1' provinceNumber={id} slotNumber={0} exists={armies[0] != null} 
+        ownsArmy={player.name == owner} onArmyClick={onArmyClick} armyObject={armies[0]} />
+      <Army key={name + '_army2'} name='army2' provinceNumber={id} slotNumber={1} exists={armies[1] != null} 
+        ownsArmy={player.name == owner} onArmyClick={onArmyClick} armyObject={armies[1]} />
+      <Army key={name + '_army3'} name='army3' provinceNumber={id} slotNumber={2} exists={armies[2] != null} 
+        ownsArmy={player.name == owner} onArmyClick={onArmyClick} armyObject={armies[2]} />
+      <Army key={name + '_army4'} name='army4' provinceNumber={id} slotNumber={3} exists={armies[3] != null} 
+        ownsArmy={player.name == owner} onArmyClick={onArmyClick} armyObject={armies[3]} />
+  
+      </div>
       
+  );
+}
+
+// Pre-defined colors to use for provinces
+const allColors = ["rgb(135, 245, 66)", "rgb(219, 78, 46)", "rgb(194, 85, 224)", 
+"rgb(82, 212, 217)", "rgb(216, 217, 167)"];
+
+/*
+  TODO: Remove
+  NOTE: Old code for province armies if anything goes wrong
+
       <button className='province_army' id='army1' style={{opacity: army1Exists}} draggable={army1Draggable} 
         onClick={() => onArmyClick(armies[0])} onDragStart={(e) => handleOnDrag(e, armies[0], id, 0)} 
         onDragEnd={(e) => handleDragEnd(e)} >Army 1</button>
@@ -79,10 +80,4 @@ export default function Province({ id, onProvinceClick, onArmyClick, name, owner
       <button className='province_army' id='army4' style={{opacity: army4Exists}} draggable={army4Draggable}
         onClick={() => onArmyClick(armies[3])} onDragStart={(e) => handleOnDrag(e, armies[3], id, 3)} 
         onDragEnd={(e) => handleDragEnd(e)} >Army 4</button>
-      </div>
-  );
-}
-
-// Pre-defined colors to use for provinces
-const allColors = ["rgb(135, 245, 66)", "rgb(219, 78, 46)", "rgb(194, 85, 224)", 
-"rgb(82, 212, 217)", "rgb(216, 217, 167)"];
+*/
